refactor(core): extract WordPiece tokenizer.json parsing
Move TokenizerFile and the normalizer / pre-tokenizer /
post-processor parsers into wordpiece/json.rs. Public
wordpiece::Tokenizer paths are unchanged.
